Add files via upload
Working version of the automatic processing of excel files and production of excitation curves. A text files with information about the processing is also procuded.
The excel files are expected to contain at least one sheet with a cell with 'Courant' and other with 'Flux'. The 5 cells below each of them are read and consdered.